Ajuste de exceções

+++ app/Http/Controllers/Api/ControllerCake.php
@@ -2,10 +2,13 @@
 
 namespace App\Http\Controllers\Api;
 
+use Exception;
+use TypeError;
 use PDOException;
 use App\Models\ModelCake;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Validator;
 
 class ControllerCake extends Controller
 {
@@ -13,23 +16,20 @@ class ControllerCake extends Controller
     /**
      * Exibe uma coleção do recurso
      *
-     * @return void
+     * @return json
      */
     public function index(){
 
         try{
             $response = ModelCake::get();
         }catch(PDOException $e){
-            $response =  ['code'=>$e->getCode(),'message'=>$e->getMessage()];
+            return response()->json(['message'=>$e->getMessage()], $e->getCode());
         }
 
-        if(is_array($response))
-            return response()->json(['message'=>$response['message']]);
-
         if(count($response)==0)
-            return response()->json(['message'=>"Nenhum bolo encontrado"]);
+            return response()->json(['message'=>"Nenhum bolo encontrado", "status"=>200]);
 
-        return response()->json($response);
+        return response()->json($response,200);
     }
 
 
@@ -44,16 +44,13 @@ class ControllerCake extends Controller
         try{
             $response = ModelCake::where('name',$name)->get();
         }catch(PDOException $e){
-            $response =  ['code'=>$e->getCode(),'message'=>$e->getMessage()];
+            return response()->json(['message'=>$e->getMessage()], $e->getCode());
         }
 
-        if(is_array($response))
-            return response()->json(['message'=>$response['message']]);
-
         if(count($response)==0)
-            return response()->json(['message'=>"Nenhum bolo encontrado para '$name'"]);
+            return response()->json(['message'=>"Nenhum bolo encontrado para '$name'"],200);
 
-        return response()->json($response);
+        return response()->json($response,200);
     }
 
 
@@ -68,7 +65,7 @@ class ControllerCake extends Controller
         try{
             $response = ModelCake::insert($request->all());
         }catch(PDOException $e){
-            $response =  ['code'=>$e->getCode(),'message'=>$e->getMessage()];
+            $response = ['code'=>$e->getCode(),'message'=>$e->getMessage()];
         }
 
         if(is_array($response)){
@@ -81,11 +78,11 @@ class ControllerCake extends Controller
                     $message = $response['message'];
                 break;
             }
-            return response()->json(['message'=>$message]);
+            return response()->json(['message'=>$message],200);
         }
 
 
-      return response()->json(['message'=>'Bolo cadastrado com sucesso']);
+      return response()->json(['message'=>'Bolo cadastrado com sucesso'],200);
     }
 
 
@@ -98,14 +95,12 @@ class ControllerCake extends Controller
      */
     public function update(Request $request, string $name){
 
-        if($request->weight == '' || $request->value == '' || $request->quantity == ''){
-            return response()->json(['message'=>"Falta informações obrigatórias (peso, valor e quantidade)"]);
-        }
+        $validator = Validator::make($request->all(), ['weight' => 'required','value' => 'required','quantity'=> 'required']);
+        if($validator->fails())
+            return response()->json(['message'=>"Falta informações obrigatórias (peso, valor e quantidade)"],200);
 
         $values = [
-            'weight'=>$request->weight,
-            'value'=>$request->value,
-            'quantity'=>$request->quantity,
+            'weight'=>$request->weight, 'value'=>$request->value, 'quantity'=>$request->quantity,
         ];
 
         try{
@@ -124,10 +119,10 @@ class ControllerCake extends Controller
                     $message = $response['message'];
                 break;
             }
-            return response()->json(['message'=>$message]);
+            return response()->json(['message'=>$message],200);
         }
 
-        return response()->json(['message'=>"Bolo '$name' atualizado com sucesso"]);
+        return response()->json(['message'=>"Bolo '$name' atualizado com sucesso"],200);
     }
 
 
@@ -139,17 +134,18 @@ class ControllerCake extends Controller
      */
     public function destroy(string $name){
         try{
-            $response =  ModelCake::where('name',$name)->delete();
+
+            $cake = ModelCake::where('name',$name)->first();
+            if(!$cake)
+                return response()->json(['message'=>"Bolo '{$name}' não existe"],200);
+
+            $response = ModelCake::where('name',$name)->delete();
         }catch(PDOException $e){
-            $response =  ['code'=>$e->getCode(),'message'=>$e->getMessage()];
+            $response = ['code'=>$e->getCode(),'message'=>$e->getMessage()];
         }
 
         if(is_array($response)){
             switch($response['code']){
-                case 404:
-                    $message = "Bolo '{$name}' não encontrado";
-                break;
-
                 case 23000:
                     $message = "Bolo '$name' não pode ser deletado pois há interesse nele";
                 break;
@@ -158,10 +154,10 @@ class ControllerCake extends Controller
                     $message = $response['message'];
                 break;
             }
-            return response()->json(['message'=>$message]);
+            return response()->json(['message'=>$message],200);
         }
 
-        return response()->json(['message'=>"Bolo '$name' removido com sucesso"]);
+        return response()->json(['message'=>"Bolo '$name' removido com sucesso"],200);
     }
 
 }
